% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV.R
\name{CalFittedSigma}
\alias{CalFittedSigma}
\title{Calculate the fitted value \eqn{A_I Cov(Z) A_I'}.}
\usage{
CalFittedSigma(Sigma, delta, Ms, arg_Ms, se_est, diagonal, merge)
}
\arguments{
\item{Sigma}{A \eqn{p} by \eqn{p} matrix.}

\item{delta}{The grid of leading constant of \eqn{\delta}.}

\item{Ms}{A vector containing the largest absolute values of each row of \code{off_Sigma}.}

\item{arg_Ms}{A vector of the indices of the largest absolute values of each
row of \code{off_Sigma}.}

\item{se_est}{The vector of the standard deviations of \eqn{p} features.}

\item{diagonal}{Logical. If TRUE, the covariance matrix of \eqn{Z} is diagonal; else FALSE.}

\item{merge}{Logical. If TRUE, take the union of all candidate pure variables;
otherwise, take the intersection.}
}
\value{
A list including: \itemize{
   \item \code{pureVec} A vector of the indices of the estimated pure variables.
   \item \code{fitted} The fitted value \eqn{A_I Cov(Z) A_I'}.
}
}
\description{
Calculate the fitted value \eqn{A_I Cov(Z) A_I'}.
}
