% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ER_estimation_LS.R
\name{Comp_ASV}
\alias{Comp_ASV}
\title{Compute the asymptotic variance of the estimated \eqn{\beta}}
\usage{
Comp_ASV(
  sigma2_hat,
  BI,
  Theta_hat,
  Gamma_hat,
  beta_hat,
  Omega_hat,
  I_hat,
  I_hat_partition
)
}
\arguments{
\item{sigma2_hat}{The estimated variance returned by \code{\link{Est_sigma2}}.}

\item{BI}{A numeric matrix.}

\item{Theta_hat}{A \eqn{p} by \eqn{K} matrix.}

\item{Gamma_hat}{The estimated diagonal elements of \eqn{Cov(E)}.}

\item{beta_hat}{The estimated \eqn{\beta}.}

\item{Omega_hat}{The estimated precision matrix of \eqn{Z}.}

\item{I_hat}{The estimated set of pure variables.}

\item{I_hat_partition}{The estimated partition of pure variables.}
}
\value{
A vector of length \eqn{K}.
}
\description{
Compute the asymptotic variance of the estimated \eqn{\beta}
}
