% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstNonpure.R
\name{Dantzig}
\alias{Dantzig}
\title{The Dantzig approach of solving one non-pure row}
\usage{
Dantzig(C_hat, y, lbd)
}
\arguments{
\item{C_hat}{A \eqn{K} by \eqn{K} matrix.}

\item{y}{A vector of length \eqn{K}.}

\item{lbd}{A tuning parameter for soft-thresholding.}
}
\value{
A vector of length \eqn{K}.
}
\description{
The Dantzig approach of solving one non-pure row
}
