% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ER.R
\name{ER}
\alias{ER}
\title{Essential Regression}
\usage{
ER(
  Y,
  X,
  res_LOVE,
  beta_est = "NULL",
  mu = 0.5,
  lbd = 0.5,
  CI = F,
  alpha_level = 0.05,
  correction = "Bonferroni"
)
}
\arguments{
\item{Y}{a vector of response.}

\item{X}{n by p data matrix.}

\item{res_LOVE}{the fitted model from \code{\link{LOVE}}.}

\item{beta_est}{the procedure of estimating beta. One of \{"NULL", "LS", "Dantzig"\}}

\item{mu}{the tuning parameter used for estimating beta via the Dantzig approach. Default value is 0.5.}

\item{lbd}{the tuning parameter used for estimating beta via the Dantzig approach. Default value is 0.5.}

\item{CI}{construct confidence intervals if TRUE, otherwise FALSE.}

\item{alpha_level}{significance level. Default set to 0.05.}

\item{correction}{correction for multiple hypothesis testing. Either NULL or "Bonferroni". Default set to "Bonferroni".}
}
\value{
a list of objects including: \itemize{
  \item \code{beta} the estimated coefficients beta
  \item \code{beta_CIs} the coordinate-wise confidence intervals of beta
  \item \code{beta_var} the variances of the estimated beta
  \item \code{coef_X} the coefficients between Y and X
  \item \code{mat_trans_to_Z} the p by K matrix used to predict Z
  \item \code{fitted_val} the fitted values
  \item \code{Z_pred} the predicted Z matrix
  \item \code{X_center} the centers of the input X
  \item \code{Y_center} the sample mean of the input Y
}
}
\description{
Perform prediction, estimation and inference under Essential Regressions
}
\examples{
data(toydata)
output_table <- LOVE(X = toydata)
}
