% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ER_estimation_Dz.R
\name{ER_est_beta_dz}
\alias{ER_est_beta_dz}
\title{Estimate \eqn{\beta} via the Dantzig approach.}
\usage{
ER_est_beta_dz(Y, X, A_hat, C_hat, I_hat, delta_opt, mu = 0.5, lbd = 0.5)
}
\arguments{
\item{Y}{A vector of response with length \eqn{n}.}

\item{X}{A \eqn{n} by \eqn{p} data matrix.}

\item{A_hat}{The estimated loading matrix.}

\item{C_hat}{The estimated covariance matrix of \eqn{Z}.}

\item{I_hat}{The estimated set of pure variables.}

\item{delta_opt}{The selected constant of \code{delta} returned by \code{\link{LOVE}}.}

\item{mu}{The tuning parameters used for estimating \eqn{\beta} via the Dantzig approach. The default value is 0.5.}

\item{lbd}{The tuning parameters used for estimating \eqn{\beta} via the Dantzig approach. The default value is 0.5.}
}
\value{
A numeric vector of length \eqn{K}.
}
\description{
Estimate \eqn{\beta} via the Dantzig approach.
}
