% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ER_estimation_LS.R
\name{ER_est_beta_LS}
\alias{ER_est_beta_LS}
\title{Estimate \eqn{\beta} via the LS approach and construct confidence intervals.}
\usage{
ER_est_beta_LS(
  Y,
  X,
  Theta_hat,
  C_hat,
  Gamma_hat,
  I_hat,
  I_hat_partition,
  BI,
  CI = T,
  alpha_level = 0.05,
  correction = NULL
)
}
\arguments{
\item{Y}{A vector of response with length \eqn{n}.}

\item{X}{A \eqn{n} by \eqn{p} data matrix.}

\item{Theta_hat}{A \eqn{p} by \eqn{K} matrix.}

\item{C_hat}{The estimated covariance matrix of \eqn{Z}.}

\item{Gamma_hat}{The estimated diagonal elements of \eqn{Cov(E)}.}

\item{I_hat}{The estimated set of pure variables.}

\item{I_hat_partition}{The estimated partition of pure variables.}

\item{BI}{A numeric matrix.}

\item{CI}{Logical. TRUE if confidence intervals are constructed.}

\item{alpha_level}{The significance level. The default set to 0.05.}

\item{correction}{Correction for addressing the multiple testing problem.
Either \code{NULL} or "Bonferroni". The default value is "Bonferroni".}
}
\description{
Estimate \eqn{\beta} via the LS approach and construct confidence intervals.
}
