% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ER_prediction.R
\name{ER_prediction}
\alias{ER_prediction}
\title{Estimation for prediction under Essential Regression}
\usage{
ER_prediction(Y, X, Theta_hat)
}
\arguments{
\item{Y}{A vector of response with length \eqn{n}.}

\item{X}{A \eqn{n} by \eqn{p} data matrix.}

\item{Theta_hat}{A \eqn{p} by \eqn{K} matrix.}
}
\value{
A list including: \itemize{
  \item \code{theta} The estimated \eqn{p}-dimensional coefficients of \eqn{X}.
  \item \code{mat_trans_to_Z} The \eqn{p} by \eqn{K} matrix used to predict \eqn{Z}.
  \item \code{fitted_val} The fitted values.
  \item \code{Z_pred} The predicted \strong{Z} matrix.
}
}
\description{
Estimate the p-dimensional coefficient between \eqn{Y} and \eqn{X}
  and the \eqn{p} by \eqn{K} matrix to predict \eqn{Z}.
}
