% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ER_estimation_LS.R
\name{Est_sigma2}
\alias{Est_sigma2}
\title{Estimate the variance of the regression error}
\usage{
Est_sigma2(Y, h, beta_hat, C_hat)
}
\arguments{
\item{Y}{A vector of response with length \eqn{n}.}

\item{h}{A numeric vector}

\item{beta_hat}{The estimated \eqn{\beta}.}

\item{C_hat}{The estimated covariance matrix of \eqn{Z}.}
}
\value{
A positive constant.
}
\description{
Estimate the variance of the regression error
}
