% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstPure.R
\name{EstAI}
\alias{EstAI}
\title{Estimate the submatrix of \eqn{A} corresponding to the pure variables.}
\usage{
EstAI(Sigma, optDelta, se_est, merge)
}
\arguments{
\item{Sigma}{A \eqn{p} by \eqn{p} matrix.}

\item{optDelta}{A numerical value.}

\item{se_est}{The vector of the standard deviations of \eqn{p} features.}

\item{merge}{Logical. If TRUE, take the union of all candidate pure variables;
otherwise, take the intersection.}
}
\value{
A list including: \itemize{
   \item \code{AI} The estimated \eqn{p} by \eqn{K} submatrix \eqn{A_I} of \eqn{A}.
   \item \code{pureVec} The vector of the indices of estimated pure variables.
   \item \code{pureSignInd} The list of the indices of estimated pure variables.
}
}
\description{
Function to calculate the fitted submatrix \eqn{A_I} and to estimate the set and
partition of the pure variables.
}
