% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstNonpure.R
\name{EstAJDant}
\alias{EstAJDant}
\title{Estimate the non-pure rows via the Dantzig approach}
\usage{
EstAJDant(C_hat, Y, lbd, se_est_J)
}
\arguments{
\item{C_hat}{A \eqn{K} by \eqn{K} matrix.}

\item{Y}{A \eqn{K} by \eqn{|J|} reponse matrix.}

\item{lbd}{A tuning parameter for soft-thresholding.}

\item{se_est_J}{The estimated standard errors of the non-pure variables.}
}
\value{
A \eqn{|J|} by \eqn{K} matrix.
}
\description{
Estimate the non-pure rows via the Dantzig approach
}
