% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstNonpure.R
\name{EstAJInv}
\alias{EstAJInv}
\title{Estimates non-pure rows via soft-thresholding}
\usage{
EstAJInv(Omega, Y, lbd)
}
\arguments{
\item{Omega}{The estimated precision matrix of \eqn{Z}.}

\item{Y}{A \eqn{K} by \eqn{|J|} reponse matrix.}

\item{lbd}{A tuning parameter for soft-thresholding.}
}
\value{
A \eqn{|J|} by \eqn{K} matrix.
}
\description{
This function estimates the \eqn{|J|} by \eqn{K} submatrix \eqn{A_J} by using soft thresholding.
}
