% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstNonpure.R
\name{EstC}
\alias{EstC}
\title{Estimate the covariance matrix of \eqn{Z}.}
\usage{
EstC(Sigma, AI, diagonal)
}
\arguments{
\item{Sigma}{A \eqn{p} by \eqn{p} matrix.}

\item{AI}{A \eqn{p} by \eqn{K} matrix.}

\item{diagonal}{Logical. If TRUE, the covariance matrix of \eqn{Z} is diagonal; else FALSE.}
}
\value{
A \eqn{K} by \eqn{K} matrix.
}
\description{
Estimate the covariance matrix of \eqn{Z}.
}
