% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstNonpure.R
\name{EstY}
\alias{EstY}
\title{Function to estimate the \eqn{K} by \eqn{|J|} submatrix of \eqn{\Sigma}.}
\usage{
EstY(Sigma, AI, pureVec)
}
\arguments{
\item{Sigma}{A \eqn{p} by \eqn{p} matrix.}

\item{AI}{A \eqn{p} by \eqn{K} matrix.}

\item{pureVec}{The estimated set of pure variables.}
}
\value{
A \eqn{K} by \eqn{|J|} matrix.
}
\description{
Function to estimate the \eqn{K} by \eqn{|J|} submatrix of \eqn{\Sigma}.
}
