% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstPure.R
\name{FindPureNode}
\alias{FindPureNode}
\title{Function to estimate the pure variables for a given \code{delta}.}
\usage{
FindPureNode(off_Sigma, delta, Ms, arg_Ms, se_est, merge)
}
\arguments{
\item{off_Sigma}{A \eqn{p} by \eqn{p} matrix.}

\item{delta}{A numeric constant.}

\item{Ms}{A vector containing the largest absolute values of each row of \code{off_Sigma}.}

\item{arg_Ms}{A vector of the indices of the largest absolute values of each
row of \code{off_Sigma}.}

\item{se_est}{The vector of the standard deviations of \eqn{p} features.}

\item{merge}{Logical. If TRUE, take the union of all candidate pure variables;
otherwise, take the intersection.}
}
\value{
A list of two objects including: \itemize{
 \item \code{pureInd} A list of the estimated indices of the pure variables.
 \item \code{pureVec} A vector of the estimated indices of the pure variables.
}
}
\description{
Function to estimate the pure variables for a given \code{delta}.
}
