% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstPure.R
\name{FindRowMax}
\alias{FindRowMax}
\title{Function to calculate the maximal absolute value for each row of the given matrix.}
\usage{
FindRowMax(Sigma)
}
\arguments{
\item{Sigma}{A \eqn{p} by \eqn{p} matrix.}
}
\value{
A numerical vector of \eqn{p} elements.
}
\description{
Function to calculate the maximal absolute value for each row of the given matrix.
}
