% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstPure.R
\name{FindRowMaxInd}
\alias{FindRowMaxInd}
\title{Find indices of a given row.}
\usage{
FindRowMaxInd(i, M, arg_M, vector, delta, se_est)
}
\arguments{
\item{i}{An integer.}

\item{M}{The maximal value of the given row.}

\item{arg_M}{The index of the maximal value.}

\item{vector}{A numeric vector.}

\item{delta}{A numerical constant.}

\item{se_est}{The vector of the standard deviations of \eqn{p} features.}
}
\value{
A vector of indices.
}
\description{
Function to calculate indices of the ith row such that the absolute
values of the corresponding entries are within \eqn{2\times}\code{delta} difference
from the given value \code{M}.
}
