% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstPure.R
\name{FindSignPureNode}
\alias{FindSignPureNode}
\title{Estimate the signs of pure variables}
\usage{
FindSignPureNode(pureList, Sigma)
}
\arguments{
\item{pureList}{A list of indices of pure variables.}

\item{Sigma}{A \eqn{p} by \eqn{p} matrix.}
}
\value{
A list of sign sub-partition of indices.
}
\description{
Function to estimate the sign sub-partition of the pure variables.
If one group has no pure variables with negative sign, then an empty list is
inserted in that position.
}
