% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOVE.R
\name{LOVE}
\alias{LOVE}
\title{Latent-model based OVErlapping clustering}
\usage{
LOVE(
  X,
  delta = seq(0.1, 1.1, 0.1),
  lbd = 0.5,
  mu = 0.5,
  merge = FALSE,
  diagonal = FALSE,
  est_non_pure_row = "HT",
  center_flag = TRUE,
  equal_var = TRUE,
  rep_CV = 50,
  verbose = TRUE
)
}
\arguments{
\item{X}{n by p data matrix.}

\item{delta}{the grid of leading constant for \eqn{\delta}. Default grid is seq(0.1, 1.1 ,0.1).}

\item{lbd}{the grid of leading constants for \eqn{\lambda}. Default value is 0.5.}

\item{mu}{the leading constant used for thresholding the loading matrix. Default value is 0.5.}

\item{merge}{if TRUE, take the union of all candidate pure variables; otherwise, take the intersection. Default set to FALSE.}

\item{diagonal}{TRUE if the covariance matrix of \eqn{Z} is diagonal; else FALSE}

\item{est_non_pure_row}{procedure used for estimating the non-pure rows. One of \{"HT", "ST", "Dantzig"\}}

\item{center_flag}{center the features if TRUE.}

\item{equal_var}{True if all features have equal variance.}

\item{rep_CV}{the number of repetitions used for cross validation.}
}
\value{
a list of objects including: \itemize{
  \item \code{K} the number of clusters.
  \item \code{pureVec} the set of pure variables.
  \item \code{pureInd} the partition of pure variables.
  \item \code{group} the clusters (indices for each cluster).
  \item \code{A} the p by K assignment matrix.
  \item \code{C} the covariance matrix of the latent variables
  \item \code{Omega} the precision of latent variables
  \item \code{Gamma} the covariance matrix of the error
}
}
\description{
Perform overlappig variable clustering of p features
collected in n samples under latent factor models
}
\examples{
data(toydata)
output_table <- LOVE(X = toydata)
}
