% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstNonpure.R
\name{LP}
\alias{LP}
\title{Cast the procedure as a linear program}
\usage{
LP(y, lbd)
}
\arguments{
\item{y}{A vector of length \eqn{K}.}

\item{lbd}{A tuning parameter for soft-thresholding.}
}
\value{
A vector of length \eqn{K}.
}
\description{
Cast the procedure as a linear program
}
