% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstPure.R
\name{Merge}
\alias{Merge}
\alias{Merge_union}
\title{Function to merge pure variables via "intersection".}
\usage{
Merge(groupList, groupVec)

Merge_union(groupList, groupVec)
}
\arguments{
\item{groupList}{An existing list of indices of pure variables.}

\item{groupVec}{A new vector of indices of pure variables.}
}
\value{
A list of indices of pure variables.
}
\description{
Function to merge pure variables via "intersection".
}
\section{Functions}{
\itemize{
\item \code{Merge_union}: Merge pure variables via "union".
}}

