% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{offSum}
\alias{offSum}
\title{Calculate the weighted sum of squares of the upper off-diagonal elements}
\usage{
offSum(M, weights)
}
\arguments{
\item{M}{A given matrix with the same dimension as \code{weights}.}

\item{weights}{A given weight matrix.}
}
\value{
A numeric value.
}
\description{
Calculate the weighted sum of squares of the upper off-diagonal elements
}
