% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ER.R
\name{Predict.ER}
\alias{Predict.ER}
\title{Prediction under Essential Regression}
\usage{
Predict.ER(fitted_ER, newX = NULL, type = "response")
}
\arguments{
\item{fitted_ER}{The object returned from \code{\link{ER}}.}

\item{newX}{A new data matrix of \eqn{X}. If not provided, the fitted values of
either \strong{Y} or \strong{Z} are returned, depending on \code{type}.}

\item{type}{Either "response" or "factor". For "response", predicted values of
\eqn{Y} are returned. For "factor", predicted \eqn{Z} is returned.}
}
\description{
Predict either the response or the latent factors under Essential Regression.
}
