% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstPure.R
\name{RecoverAI}
\alias{RecoverAI}
\title{Function to return the estimated submatrix \eqn{A_I} from the partition of
pure variables.}
\usage{
RecoverAI(estGroupList, p)
}
\arguments{
\item{estGroupList}{A list of group indices of the pure variables.}
}
\value{
A \eqn{p} by \eqn{K} matrix.
}
\description{
Function to return the estimated submatrix \eqn{A_I} from the partition of
pure variables.
}
