% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{recoverGroup}
\alias{recoverGroup}
\title{Recover clusters based on a given \eqn{p} by \eqn{K} loading matrix.}
\usage{
recoverGroup(A)
}
\arguments{
\item{A}{A \eqn{p} by \eqn{K} matrix.}
}
\value{
A list of group indices with sign sub-partition.
}
\description{
Recover clusters based on a given \eqn{p} by \eqn{K} loading matrix.
}
