% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{singleton}
\alias{singleton}
\title{Function to check if there exists any element in the given list that has length
equal to 1}
\usage{
singleton(estPureIndices)
}
\arguments{
\item{estPureIndices}{A list of indices of the estimated pure variables.}
}
\value{
Logical. If exists at least one, return TRUE; otherwise return FALSE
}
\description{
Function to check if there exists any element in the given list that has length
equal to 1
}
