% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ER_estimation_Dz.R
\name{Solve_dz}
\alias{Solve_dz}
\title{Solve the linear program to estimate \eqn{\beta}.}
\usage{
Solve_dz(C, h, mu, lbd)
}
\arguments{
\item{C}{A \eqn{K} by \eqn{K} matrix.}

\item{h}{A vector of length \eqn{K}.}

\item{mu, lbd}{Two numeric constants.}
}
\description{
Solve the linear program to estimate \eqn{\beta}.
}
