% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstOmega.R
\name{solve_row}
\alias{solve_row}
\title{Estimate each row by solving a LP}
\usage{
solve_row(col_ind, C_hat, lbd)
}
\arguments{
\item{col_ind}{An integer.}

\item{C_hat}{A \eqn{K} by \eqn{K} matrix.}

\item{lbd}{A numeric constant.}
}
\value{
A vector of length \eqn{K}.
}
\description{
Estimate each row by solving a LP
}
