% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstPure.R
\name{TestPure}
\alias{TestPure}
\title{Test pure variable.}
\usage{
TestPure(Sigma_row, rowInd, Si, Ms, arg_Ms, delta, se_est)
}
\arguments{
\item{Sigma_row}{A given row of \code{Sigma}.}

\item{rowInd}{An integer index.}

\item{Si}{A vector of indices.}

\item{Ms}{A vector containing the largest absolute values of each row of \code{off_Sigma}.}

\item{arg_Ms}{A vector of the indices of the largest absolute values of each
row of \code{off_Sigma}.}

\item{delta}{A numeric constant.}

\item{se_est}{The vector of the standard deviations of \eqn{p} features.}
}
\value{
Logical. TRUE or FALSE.
}
\description{
Function to check if a given row corresponds to a pure variable.
}
