% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{threshA}
\alias{threshA}
\title{Function to hard-threshold}
\usage{
threshA(A, mu, scale = FALSE)
}
\arguments{
\item{A}{A \eqn{p} by \eqn{K} matrix.}

\item{mu}{A numeric value.}

\item{scake}{Logical. Normalize the row-wise \eqn{\ell-1} norm if TRUE.}
}
\value{
A matrix with the same dimension as \code{A}.
}
\description{
Threshold the estimated \code{A} based on the given \code{mu}.
  If \code{scale} is TRUE, then normalize each row of \code{A} such that the
  \eqn{\ell-1} norm of each row is no larger than 1.
}
