% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ER.R
\name{ER}
\alias{ER}
\title{Essential Regression}
\usage{
ER(
  Y,
  X,
  res_LOVE,
  beta_est = "LS",
  mu = 0.5,
  lbd = 0.5,
  CI = F,
  alpha_level = 0.05,
  correction = "Bonferroni"
)
}
\arguments{
\item{Y}{A vector of response with length \eqn{n}.}

\item{X}{A \eqn{n} by \eqn{p} data matrix.}

\item{res_LOVE}{The returned object from \code{\link{LOVE}}.}

\item{beta_est}{The procedure used for estimating \eqn{\beta}. One of \{\code{NULL}, "LS", "Dantzig"\}}

\item{mu, lbd}{The tuning parameters used for estimating \eqn{\beta} via the Dantzig approach. The default value is 0.5.}

\item{CI}{Logical. TRUE if confidence intervals are constructed.}

\item{alpha_level}{The significance level. The default set to 0.05.}

\item{correction}{The approach used for addressing the multiple testing problem.
Either \code{NULL} or "Bonferroni". The default value is "Bonferroni".}
}
\value{
A list of objects including: \itemize{
  \item \code{beta} The estimated coefficients of \eqn{\beta}.
  \item \code{beta_CIs} The coordinate-wise confidence intervals of \eqn{\beta}.
  \item \code{beta_var} The variances of the \code{beta}.
  \item \code{coef_X} The estimated p-dimensional coefficient between \eqn{Y} and \eqn{X}.
  \item \code{mat_trans_to_Z} The \eqn{p} by \eqn{K} matrix used to predict \eqn{Z}.
  \item \code{fitted_val} The fitted values of length \eqn{n}.
  \item \code{Z_pred} The predicted \strong{Z} matrix.
  \item \code{X_center} Centers of the input \code{X}.
  \item \code{Y_center} Center of the input \code{Y}.
}
}
\description{
Under the Essential
 Regression framework \deqn{X = AZ+E,    Y = Z' \beta + \epsilon,}
 perform prediction of \eqn{Y}, estimation and inference of \eqn{\beta}.
}
\examples{
data(toydata)
output_table <- ER(X = toydata)
}
