% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstOmega.R
\name{estOmega}
\alias{estOmega}
\title{Estimate the precision matrix of \eqn{Z} via LP}
\usage{
estOmega(lbd, C_hat)
}
\arguments{
\item{lbd}{A numeric constant.}

\item{C_hat}{A \eqn{K} by \eqn{K} matrix.}
}
\value{
A \eqn{K} by \eqn{K} matrix.
}
\description{
Estimate the precision matrix of \eqn{Z} via LP
}
