% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOVE.R
\name{LOVE}
\alias{LOVE}
\title{LOVE: Latent-model based OVErlapping clustering}
\usage{
LOVE(
  X,
  delta = seq(0.1, 1.1, 0.1),
  lbd = 0.5,
  mu = 0.5,
  merge = FALSE,
  diagonal = FALSE,
  est_non_pure_row = "HT",
  center_flag = TRUE,
  equal_var = TRUE,
  rep_CV = 50,
  verbose = TRUE
)
}
\arguments{
\item{X}{A \eqn{n} by \eqn{p} data matrix.}

\item{delta}{The grid of leading constant of \eqn{\delta}.}

\item{lbd}{The grid of leading constant of \eqn{\lambda}.}

\item{mu}{The leading constant used for thresholding the loading matrix.}

\item{merge}{Logical. If TRUE, take the union of all candidate pure variables;
otherwise, take the intersection.}

\item{diagonal}{Logical. If TRUE, the covariance matrix of \eqn{Z} is diagonal; else FALSE.}

\item{est_non_pure_row}{String. Procedure used for estimating the non-pure rows. One of \{"HT", "ST", "Dantzig"\}.}

\item{center_flag}{Logical. If TRUE, center the features.}

\item{equal_var}{Logical. TRUE if all features have equal variance.}

\item{rep_CV}{The number of repetitions used for cross validation.}

\item{verbose}{Logical. Set FALSE to suppress printing the progress.}
}
\value{
A list of objects including: \itemize{
  \item \code{K} The estimated number of clusters.
  \item \code{pureVec} The estimated set of pure variables.
  \item \code{pureInd} The estimated partition of pure variables.
  \item \code{group} The estimated clusters (indices of each cluster).
  \item \code{A} The estimated \eqn{p} by \eqn{K} assignment matrix.
  \item \code{C} The covariance matrix of \eqn{Z}.
  \item \code{Omega} The precision matrix of \eqn{Z}.
  \item \code{Gamma} The diagonal of the covariance matrix of \eqn{E}.
  \item \code{optDelta} The selected value from \code{delta}.
}
}
\description{
Perform overlapping (variable) clustering of a \eqn{p}-dimensional feature
generated from the latent factor model \deqn{X = AZ + E} with identifiability
conditions on \eqn{A} and \eqn{Cov(Z)}.
}
\examples{
p <- 6
n <- 50
K <- 2
A <- rbind(c(1, 0), c(-1, 0), c(0, 1), c(0, 1), c(1/3, 2/3), c(1/2, -1/2))
Z <- matrix(rnorm(n * K, sd = 2), n, K)
E <- matrix(rnorm(n * p), n, p)
X <- Z \%*\% t(A) + E
res_LOVE <- LOVE(X)
}
