% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{offSum}
\alias{offSum}
\title{Calculate the weighted sum of squares of the upper off-diagonal elements}
\usage{
offSum(M, weights)
}
\arguments{
\item{M}{A given symmetric matrix}

\item{weights}{A vector of length equal to the number of rows of \code{M}.}
}
\value{
A numeric value.
}
\description{
Calculate the weighted sum of squares of the upper off-diagonal elements
}
